% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-manipulation.R
\name{add_article}
\alias{add_article}
\title{Append a/an to word}
\usage{
add_article(x)
}
\arguments{
\item{x}{string or numeric value}
}
\description{
Append a/an to word
}
\examples{
add_article("increase")
add_article("decrease")
}
