% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-defaults.R
\name{augment_article_patterns}
\alias{augment_article_patterns}
\title{Adjust which words start with "a" vs "an"}
\usage{
augment_article_patterns(regex_for_a, regex_for_an)
}
\arguments{
\item{regex_for_a}{regular expression for words that get the article 'a'}

\item{regex_for_an}{regular expression for words that get the article 'an'}
}
\description{
headliner uses crude logic to anticipate the articles used before phrases
like "(an) increase", "(a) decrease", "(an) 83" and "(a) -5"
}
\details{
These arguments assume that the words start with the patterns you provide.
Passing "hour|heir" will be translated to "^(hour|heir)".
}
\examples{
# The crude logic for headliner would give all "h" words the article "a"
# by default, the word "heirloom" returns "a"
add_article("heirloom")

# the patterns that are used can be updated
augment_article_patterns(
  regex_for_a = "Euro|uni",
  regex_for_an = "hour|heir"
)

add_article("heirloom")

# these patterns can be reset using 'NULL'
augment_article_patterns(
  regex_for_an = NULL
)

add_article("heirloom")
}
\seealso{
\code{\link[=get_headliner_defaults]{get_headliner_defaults()}}, \code{\link[=set_headliner_defaults]{set_headliner_defaults()}}
}
