% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_hd}
\alias{build_hd}
\title{Build headline pages}
\usage{
build_hd(.data, save_as = "headlines.html", title = "Headlines")
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble),
or a lazy data frame.  Expects the following string columns
\emph{title} (required), \emph{description}, \emph{link}, \emph{pubDate}, and \emph{image}.}

\item{save_as}{A string with the name of the static page for the output.}

\item{title}{A title of the headlines page}
}
\description{
Build headline pages
}
