% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-phrase.R
\name{build_phrase}
\alias{build_phrase}
\alias{build_phrase.default}
\alias{build_phrase.list}
\title{Title}
\usage{
build_phrase(...)

\method{build_phrase}{default}(
  compare,
  reference,
  calc = c("value", "prop"),
  phrasing = headliner::phrase_terms()
)

\method{build_phrase}{list}(x, compare, reference, ...)
}
\arguments{
\item{...}{}

\item{compare}{numeric value to compare against reference (base) value}

\item{reference}{numeric value that 'compare' value will be compared against}

\item{calc}{string should comparison be made as the difference between the
two ('value', y - x) or the percent difference ('prop', (y - x) / x)}

\item{phrasing}{list of values to use for when y is more than x, y is the
same as x, or y is less than x.}

\item{x}{a named list with values to compare}
}
\description{
Title

Build phrase components
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Build phrase components from named list
}}

\examples{
build_phrase(10, 8) \%>\% head(2)
build_phrase(10, 8, calc = "prop") \%>\% head(2)
build_phrase(10, 8, phrasing = phrase_terms(more = "higher")) \%>\% head(2)
build_phrase(list(a = 1, b = 2), a, b) \%>\% head(2)
}
\seealso{
\code{\link[=view_components]{view_components()}} and \code{\link[=phrase_terms]{phrase_terms()}}
}
