% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_simple}
\alias{build_simple}
\title{Build a simple headline page}
\usage{
build_simple(x, saveas = "headlines.html", title = "Title")
}
\arguments{
\item{x}{A dataframe containing columns title, description, and link.}

\item{saveas}{A string with the name of the static page for the output.}

\item{title}{A title of the headlines page}
}
\description{
Build a simple headline page
}
