% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-columns.R
\name{compare_columns}
\alias{compare_columns}
\title{Compare two columns within a data frame}
\usage{
compare_columns(df, cols = everything(), calc = list(mean = mean))
}
\arguments{
\item{df}{data frame}

\item{cols}{columns to compare, supports select helpers like 'starts_with()'
or 'where(is.numeric)'}

\item{calc}{named list of the functions to use, ex:
list(mean = mean, sd = sd) 'purrr' style phrases are also supported like
list(mean = ~mean(.x, na.rm = TRUE), sd = sd)}
}
\description{
Compare two columns within a data frame
}
\examples{
# by default the function will compute the means
compare_columns(flights_jfk, c("dep_delay", "arr_delay"))

# you can name the output using 'list()' and select the columns using
# any of the 'select()' helpers
compare_columns(flights_jfk, ends_with("delay"), list(avg = mean))
}
