% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-data.R
\name{demo_data}
\alias{demo_data}
\title{Small dataset referencing the current date}
\usage{
demo_data(n = 10, by = "-2 month")
}
\arguments{
\item{n}{number of rows to return}

\item{by}{string indicating the unit of time between dates in
\code{seq.Date(..., by = )}}
}
\description{
Small dataset referencing the current date
}
\examples{
demo_data()
}
