% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flights_jfk}
\alias{flights_jfk}
\title{Flights from JFK Airport}
\format{
A tibble with 12,367 rows and 10 columns:
\describe{
\item{year}{year of arrival, 2012-2013}
\item{date}{scheduled date of arrival}
\item{hour}{hour of scheduled arrival, 9 AM to 9 PM}
\item{carrier}{airline carrier, American Airlines (AA), Jet Blue (B6),
and Delta (DL)}
\item{dest}{destination, Las Vegas (LAS), Los Angeles (LAX),
and San Francisco (SFO)}
\item{tailnum}{tail number of plane}
\item{distance}{distance from origin to destination}
\item{air_time}{time in air from origin to destination}
\item{dep_delay}{departure delay in minutes}
\item{arr_delay}{arrival delay in minutes}
}
}
\usage{
flights_jfk
}
\description{
A modified subset of  \code{\link[nycflights13]{flights}}.
}
\keyword{datasets}
