% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-manipulation.R
\name{get_article}
\alias{get_article}
\title{Choose "a" or "an"
Definition listed under \code{\link[=add_article]{add_article()}}}
\usage{
get_article(x)
}
\arguments{
\item{x}{a number or string}
}
\description{
Choose "a" or "an"
Definition listed under \code{\link[=add_article]{add_article()}}
}
\examples{
get_article("increase")
get_article("decrease")
get_article(5)
get_article(8)
get_article(18123)
stats::setNames(
  get_article(1.8 * 10^(1:7)),
  prettyNum(1.8 * 10^(1:7), big.mark = ",")
)

}
