% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-defaults.R
\name{get_headliner_defaults}
\alias{get_headliner_defaults}
\title{List the defaults that headliner is using}
\usage{
get_headliner_defaults()
}
\description{
The default phrases used by glue might not be easy to remember. This
function will show the values headliner uses as default.
}
\examples{
get_headliner_defaults()
}
\seealso{
\code{\link[=augment_article_patterns]{augment_article_patterns()}}, \code{\link[=set_headliner_defaults]{set_headliner_defaults()}}
}
