% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headline.R
\name{headline.list}
\alias{headline.list}
\alias{headline.data.frame}
\title{For a list}
\usage{
\method{headline}{list}(x, compare, reference, ...)

\method{headline}{data.frame}(x, compare, reference, ...)
}
\arguments{
\item{x}{data frame, must be a single row}

\item{compare}{numeric value to compare against reference (base) value}

\item{reference}{numeric value that 'compare' value will be compared against}

\item{...}{
  Arguments passed on to \code{\link[=compare_values]{compare_values}}
  \describe{
    \item{\code{trend_phrasing}}{list of values to use for when y is more than x, y is the
same as x, or y is less than x.}
    \item{\code{plural_phrases}}{named list of values to use when difference (delta) is
singular (delta = 1) or plural (delta != 1)}
    \item{\code{orig_values}}{a string using \code{\link[glue]{glue}} syntax. \code{{c}} =
the 'compare' value, and \code{{r}} = 'reference'}
    \item{\code{n_decimal}}{numeric value to limit the number of decimal places in
the returned values.}
    \item{\code{round_all}}{logical value to indicate if all values should be rounded.
When FALSE, the values will return with no modification. When TRUE (default)
all values will be round to the length specified by 'n_decimal'.}
    \item{\code{scale}}{number indicating the scaling factor. When scale = 1, 1/4 will
return 0.25, when scale = 100 (default) 1/4 will return 25}
  }}
}
\description{
For a list

For a data frame
}
\section{Functions}{
\itemize{
\item \code{headline.list}: for lists

\item \code{headline.data.frame}: for data frames
}}

