% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{headlines}
\alias{headlines}
\title{Build headline pages}
\usage{
headlines(.data, save_as, title = "Headlines", layout = "list")
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble),
or a lazy data frame.  Expects the following string columns
\emph{title} (required), \emph{description}, \emph{link}, \emph{pub_date}, and \emph{thumbnail}.}

\item{save_as}{A string with the output file name.
Either \emph{.pdf} and \emph{.html}.}

\item{title}{A string with the title of the headlines page.}

\item{layout}{A string with the style layout. Pass "list" or "card".}
}
\value{
0
}
\description{
Build headline pages
}
\examples{
\dontrun{
data(sciencegeist)
sciencegeist <- utils::head(sciencegeist)
headlines(sciencegeist, "sciencegeist.html")
}
}
