% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-phrase.R
\name{phrase_terms}
\alias{phrase_terms}
\title{Phrases for direction of difference}
\usage{
phrase_terms(more = "increase", less = "decrease", same = "difference")
}
\arguments{
\item{more}{string to use when x > y}

\item{less}{string to use when x < y}

\item{same}{string to use when x == y}
}
\description{
Phrases for direction of difference
}
\examples{
phrase_terms(same = "no change")
}
\seealso{
\code{\link[=build_phrase]{build_phrase()}}
}
