% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plural-phrasing.R
\name{plural_phrasing}
\alias{plural_phrasing}
\title{List of values to use when change is plural (or singular)}
\usage{
plural_phrasing(single, multi)
}
\arguments{
\item{single}{string to use when delta = 1}

\item{multi}{string to use when delta > 1}
}
\description{
List of values to use when change is plural (or singular)
}
\examples{
plural_phrasing(single = "person", multi = "people")

headline(
  x = 1:2,
  y = 0,
  headline = "a difference of {delta} {people}",
  plural_phrases = list(people = plural_phrasing("person", "people"))
)
}
