% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{rss_to_df}
\alias{rss_to_df}
\title{Parse an RSS feed (XML)}
\usage{
rss_to_df(rss_feed)
}
\arguments{
\item{rss_feed}{A string pointing to the RSS URL.}
}
\value{
A tibble with RSS data.
}
\description{
Parse an RSS feed (XML)
}
\examples{
rss_to_df("https://www.sciencegeist.ch/news/rss")
}
