% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sciencegeist}
\alias{sciencegeist}
\title{Headlines sourced from sciencegeist.com's RSS feed}
\format{
A data frame with 118 rows and 4 columns:
\describe{
\item{title}{Title of the headline on sciencegeist.}
\item{description}{Description of the headline.}
\item{pubDate}{Date of publication on sciencegeist.}
\item{link}{Link to the original source.}
\item{image}{Link to image on sciencegeist.}
}
}
\source{
\url{https://www.sciencegeist.ch/news/rss}
}
\usage{
sciencegeist
}
\description{
A dataset containing 118 headlines in science policy.
}
\keyword{datasets}
