% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-defaults.R
\name{set_headliner_defaults}
\alias{set_headliner_defaults}
\title{Set the default values used in making headlines}
\usage{
set_headliner_defaults(
  headline = NULL,
  orig_values = NULL,
  if_match = NULL,
  plural_phrases = NULL,
  n_decimal = NULL,
  round_all = NULL,
  multiplier = NULL
)
}
\arguments{
\item{orig_values}{a string using \code{\link[glue]{glue}} syntax.
example: \verb{(\{x\} vs \{y\})}}

\item{plural_phrases}{named list of values to use when difference (delta) is
singular (delta = 1) or plural (delta != 1)}

\item{n_decimal}{numeric value to limit the number of decimal places in
the returned values.}

\item{round_all}{logical value to indicate if all values should be rounded.
When FALSE, the values will return with no modification. When TRUE (default)
all values will be round to the length specified by 'n_decimal'.}

\item{multiplier}{number indicating the scaling factor. When multiplier = 1
(default), 0.25 will return 0.25. When multiplier = 100, 0.25 will return 25.}
}
\description{
Set the default values used in making headlines
}
\examples{
# the way headliner writes out-of-the-box
headline(3, 4)

# the values headliner uses under the hood
get_headliner_defaults()

# you can use compare_values(2, 3) |> view_list() to see all talking
# points but we can use this as a short cut to see some of them:
names(compare_values(2, 3)) |> head()
# update headline
set_headliner_defaults(
  headline = "{article_delta_p}\% {trend} ({orig_values})",
  orig_values = "from {y} to {x}"
)

# view changes
get_headliner_defaults()

# and in action
headline(3, 4)

# reset values
set_headliner_defaults(headline = NULL, orig_values = NULL)
headline(3, 4)
}
\seealso{
\code{\link[=augment_article_patterns]{augment_article_patterns()}}, \code{\link[=get_headliner_defaults]{get_headliner_defaults()}}
}
