% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend-terms.R
\name{trend_terms}
\alias{trend_terms}
\title{Phrases for direction of difference}
\usage{
trend_terms(more = "increase", less = "decrease", same = "difference")
}
\arguments{
\item{more}{string to use when x > y}

\item{less}{string to use when x < y}

\item{same}{string to use when x == y}
}
\description{
Phrases for direction of difference
}
\examples{
trend_terms(same = "no change")

headline(
  x = c(9, 11),
  y = 10,
  headline = "{trend} by {delta_p}\%",
  trend_phrases = trend_terms("higher", "lower")
)

}
\seealso{
\code{\link[=compare_values]{compare_values()}}
}
