% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-phrase.R
\name{view_components}
\alias{view_components}
\title{Compact view of phrase components}
\usage{
view_components(x)
}
\arguments{
\item{x}{list from 'build_phrase()'}
}
\description{
Compact view of phrase components
}
\examples{
build_phrase(10, 8) \%>\% view_components()
}
\seealso{
\code{\link[=build_phrase]{build_phrase()}}
}
