% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-phrase.R
\name{view_list}
\alias{view_list}
\title{Compact view of list values}
\usage{
view_list(x)
}
\arguments{
\item{x}{list from 'compare_values()'}
}
\description{
Compact view of list values
}
\examples{
compare_values(10, 8) \%>\%
  view_list()
}
\seealso{
\code{\link[=compare_values]{compare_values()}}
}
