% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-phrase.R
\name{build_phrase.default}
\alias{build_phrase.default}
\alias{build_phrase.list}
\title{Build phrase components}
\usage{
\method{build_phrase}{default}(
  compare,
  reference,
  calc = c("value", "prop"),
  phrasing = headliner::phrase_terms(),
  expr = "{compare} vs. {reference}"
)

\method{build_phrase}{list}(x, compare, reference, ...)
}
\arguments{
\item{compare}{numeric value to compare against reference (base) value}

\item{reference}{numeric value that 'compare' value will be compared against}

\item{calc}{string should comparison be made as the difference between the
two ('value', y - x) or the percent difference ('prop', (y - x) / x)}

\item{phrasing}{list of values to use for when y is more than x, y is the
same as x, or y is less than x.}

\item{expr}{a string using \code{\link[glue]{glue}} syntax}

\item{x}{a named list with values to compare}
}
\description{
Build phrase components
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Build phrase components from named list
}}

\examples{
# manually entered

build_phrase(10, 8) \%>\% head(2)
build_phrase(10, 8, calc = "prop") \%>\% head(2)
build_phrase(10, 8, phrasing = phrase_terms(more = "higher")) \%>\% head(2)

# a phrase about the comparion can be edited by providing glue syntax
build_phrase(10, 8, expr = "{compare} to {reference} people")$expr

# Using a list as will be output of compare_conditions()

# First a simplified example
list(a = 1, b = 2) \%>\%
  build_phrase(a, b) \%>\%
  head(2)

# How it is used with compare_conditions()
res <-
  flights_jfk \%>\%
  compare_conditions(
    compare = carrier == "AA",
    reference = carrier == "DL",
    arr_delay
  )

res

res \%>\%
  lapply(round, 1) \%>\% # prevents endless decimal places
  build_phrase(
    comp_arr_delay_mean,
    ref_arr_delay_mean
  ) \%>\%
  head(2)
}
\seealso{
\code{\link[=view_components]{view_components()}} and \code{\link[=phrase_terms]{phrase_terms()}}
}
