% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-groups.R
\name{compare_columns}
\alias{compare_columns}
\title{Compare two columns within a data frame}
\usage{
compare_columns(df, ..., calc = list(mean = mean))
}
\arguments{
\item{df}{data frame}

\item{...}{columns to compare, supports select helpers like 'starts_with()'
or 'where(is.numeric)'}

\item{calc}{named list of the functions to use, ex:
list(mean = mean, sd = sd) 'purrr' style phrases are also supported like
list(mean = ~mean(.x, na.rm = TRUE), sd = sd)}
}
\description{
Compare two columns within a data frame
}
\examples{
compare_columns(flights_jfk, )
compare_columns(flights_jfk, ends_with("delay"))
}
