% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-conditions.R
\name{compare_conditions}
\alias{compare_conditions}
\title{Compare two conditions within a data frame}
\usage{
compare_conditions(df, x, y, cols = everything(), calc = list(mean = mean))
}
\arguments{
\item{df}{data frame}

\item{x}{condition for comparison, same criteria you would use in
'dplyr::filter', used in contrast to the reference group 'y'}

\item{y}{condition for comparison, same criteria you would use in
'dplyr::filter', used in contrast to the reference group 'x'}

\item{cols}{columns to use in comparison}

\item{calc}{named list of the functions to use, ex:
list(mean = mean, sd = sd) 'purrr' style phrases are also supported like
list(mean = ~mean(.x, na.rm = TRUE), sd = sd)}
}
\description{
Compare two conditions within a data frame
}
\examples{
# compare_conditions() produces a list that can be passed to headline_list()
flights_jfk |>
  compare_conditions(
    x = (hour > 12),
    y = (hour <= 12),
    dep_delay
  )

.Last.value |> headline_list()

 # you can return multiple objects to compare
 # 'view_List()' is a helper to see list objects in a compact way
flights_jfk |>
  compare_conditions(
    x = (hour > 12),
    y = (hour <= 12),
    cols = c(dep_delay, arr_delay),
    calc = list(mean = mean, sd = sd)
  ) |>
  view_list()

# if you want to compare to the average use 'complete.cases(.)'
flights_jfk |>
  compare_conditions(
    x = (hour > 12),
    y = complete.cases(.),
    dep_delay
  ) |>
  view_list()
}
