% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-groups.R
\name{compare_conditions}
\alias{compare_conditions}
\title{Title}
\usage{
compare_conditions(
  df,
  compare = complete.cases(.),
  reference = complete.cases(.),
  ...,
  calc = list(mean = mean),
  cond = NULL
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
