% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-columns.R, R/compare-conditions.R
\name{compare_columns}
\alias{compare_columns}
\alias{compare_conditions}
\title{Compare two columns within a data frame}
\usage{
compare_columns(df, cols = everything(), calc = list(mean = mean))

compare_conditions(
  df,
  compare,
  reference,
  cols = everything(),
  calc = list(mean = mean)
)
}
\arguments{
\item{df}{data frame}

\item{cols}{columns to use in comparison}

\item{calc}{named list of the functions to use, ex:
list(mean = mean, sd = sd) 'purrr' style phrases are also supported like
list(mean = ~mean(.x, na.rm = TRUE), sd = sd)}

\item{compare}{condition for comparison, same criteria you would use in
'dplyr::filter'}

\item{reference}{same as compare}
}
\description{
Compare two columns within a data frame

Compare two conditions within a data frame
}
\examples{
# by default the function will compute the means
compare_columns(flights_jfk, c("dep_delay", "arr_delay"))

# you can name the output using 'list()' and select the columns using
# any of the 'select()' helpers
compare_columns(flights_jfk, ends_with("delay"), list(avg = mean))
# compare_conditions() produces a list that can be passed to headline_list()
flights_jfk \%>\%
  compare_conditions(
    compare = (hour > 12),
    reference = (hour <= 12),
    dep_delay
  )

.Last.value \%>\% headline_list()

 # you can return multiple objects to compare
 # 'view_List()' is a helper to see list objects in a compact way
flights_jfk \%>\%
  compare_conditions(
    compare = (hour > 12),
    reference = (hour <= 12),
    cols = c(dep_delay, arr_delay),
    calc = list(mean = mean, sd = sd)
  ) \%>\%
  view_list()

# if you want to compare to the average use 'complete.cases(.)'
flights_jfk \%>\%
  compare_conditions(
    compare = (hour > 12),
    reference = complete.cases(.),
    dep_delay
  ) \%>\%
  view_list()
}
