% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-phrase.R
\name{compare_values.default}
\alias{compare_values.default}
\alias{compare_values.list}
\title{Compare two values and get talking points}
\usage{
\method{compare_values}{default}(
  compare,
  reference,
  calc = c("value", "prop"),
  trend_phrasing = headliner::trend_terms(),
  expr = "{c} vs. {r}",
  n_decimal = 1,
  round_all = TRUE,
  scale = 100
)

\method{compare_values}{list}(x, compare, reference, ...)
}
\arguments{
\item{compare}{numeric value to compare against reference (base) value}

\item{reference}{numeric value that 'compare' value will be compared against}

\item{calc}{string should comparison be made as the difference between the
two ('value', y - x) or the percent difference ('prop', (y - x) / x)}

\item{trend_phrasing}{list of values to use for when y is more than x, y is the
same as x, or y is less than x.}

\item{expr}{a string using \code{\link[glue]{glue}} syntax. \code{{c}} =
the 'compare' value, and \code{{r}} = 'reference'}

\item{n_decimal}{numeric value to limit the number of decimal places in
the returned values.}

\item{round_all}{logical value to indicate if all values should be rounded.
When FALSE, the values will return with no modification. When TRUE (default)
all values will be round to the length specified by 'n_decimal'.}

\item{scale}{number indicating the scaling factor. When scale = 1, 1/4 will
return 0.25, when scale = 100 (default) 1/4 will return 25}

\item{x}{a named list with values to compare}
}
\description{
Compare two values and get talking points
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Build phrase components from named list
}}

\examples{
# manually entered

compare_values(10, 8) \%>\% head(2)
compare_values(10, 8, calc = "prop") \%>\% head(2)
compare_values(10, 8, trend_phrasing = trend_terms(more = "higher")) \%>\%
  head(2)

# a phrase about the comparion can be edited by providing glue syntax
# 'c' = the 'compare' value, 'r' = 'reference'
compare_values(10, 8, expr = "{c} to {r} people")$expr

# you can also adjust the rounding, although the default is 1
compare_values(22/7, 22/3)$expr
compare_values(22/7, 22/3, n_decimal = 3)$expr

# Using a list as will be output of compare_conditions()

# First a simplified example
list(a = 1, b = 2) \%>\%
  compare_values(a, b) \%>\%
  head(2)

# How it is used with compare_conditions()
res <-
  flights_jfk \%>\%
  compare_conditions(
    compare = carrier == "AA",
    reference = carrier == "DL",
    arr_delay
  )

res

res \%>\%
  lapply(round, 1) \%>\% # prevents endless decimal places
  compare_values(
    comp_arr_delay_mean,
    ref_arr_delay_mean
  ) \%>\%
  head(2)
}
\seealso{
\code{\link[=view_list]{view_list()}} and \code{\link[=trend_terms]{trend_terms()}}
}
