% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-values.R
\name{compare_values}
\alias{compare_values}
\title{Compare two values and get talking points}
\usage{
compare_values(
  compare,
  reference,
  trend_phrasing = headliner::trend_terms(),
  orig_values = "{c} vs. {r}",
  plural_phrases = NULL,
  n_decimal = 1,
  round_all = TRUE,
  scale = 100
)
}
\arguments{
\item{compare}{a numeric value to compare to a reference value}

\item{reference}{a numeric value to act as a control for the 'compare' value}

\item{trend_phrasing}{list of values to use for when y is more than x, y is the
same as x, or y is less than x.}

\item{orig_values}{a string using \code{\link[glue]{glue}} syntax. \code{{c}} =
the 'compare' value, and \code{{r}} = 'reference'}

\item{plural_phrases}{named list of values to use when difference (delta) is
singular (delta = 1) or plural (delta != 1)}

\item{n_decimal}{numeric value to limit the number of decimal places in
the returned values.}

\item{round_all}{logical value to indicate if all values should be rounded.
When FALSE, the values will return with no modification. When TRUE (default)
all values will be round to the length specified by 'n_decimal'.}

\item{scale}{number indicating the scaling factor. When scale = 1, 1/4 will
return 0.25, when scale = 100 (default) 1/4 will return 25}
}
\description{
Compare two values and get talking points
}
\examples{
# the values can be manually entered

compare_values(c(10, 8)) \%>\% head(2)
# percent difference (10-8)/8
compare_values(c(10, 8))$delta_p
compare_values(c(10, 8), trend_phrasing = trend_terms(more = "higher")) \%>\%
  head(2)

# a phrase about the comparion can be edited by providing glue syntax
# 'c' = the 'compare' value, 'r' = 'reference'
compare_values(c(10, 8), orig_values = "{c} to {r} people")$orig_values

# you can also adjust the rounding, although the default is 1
compare_values(c(0.1234, 0.4321))$orig_values
compare_values(c(0.1234, 0.4321), n_decimal = 3)$orig_values
}
\seealso{
\code{\link[=view_list]{view_list()}}, \code{\link[=trend_terms]{trend_terms()}}, and \code{\link[=plural_phrasing]{plural_phrasing()}}
}
