% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-data.R
\name{demo_data}
\alias{demo_data}
\title{Small data set referencing the current date}
\usage{
demo_data(n = 10, by = "-2 month")
}
\arguments{
\item{n}{number of rows to return}

\item{by}{string indicating the unit of time between dates in
\code{seq.Date(..., by = )}}
}
\value{
Returns a data frame of size \code{n}.
}
\description{
Small data set referencing the current date
}
\examples{
demo_data()

demo_data(n = 8, by = "1 day")
}
