% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-manipulation.R
\name{get_article}
\alias{get_article}
\title{Choose "a" or "an"}
\usage{
get_article(x)
}
\arguments{
\item{x}{a number or string}
}
\description{
Choose "a" or "an"
}
\examples{
get_article("increase")
get_article("decrease")
get_article(5)
get_article(8)
}
