% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headline.R
\name{headline}
\alias{headline}
\alias{headline.default}
\alias{headline.list}
\title{Compose phrases that describe differences in the data}
\usage{
headline(...)

\method{headline}{default}(
  compare,
  reference,
  headline = "{trend} of {delta} ({orig_values})",
  ...,
  if_match = "There was no difference.",
  trend_phrases = headliner::trend_terms(),
  plural_phrases = NULL,
  orig_values = "{c} vs. {r}",
  n_decimal = 1,
  round_all = TRUE,
  multiplier = 1,
  return_data = FALSE
)

\method{headline}{list}(x, compare, reference, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=headline.default]{headline.default}}
  \describe{
    \item{\code{}}{}
  }}

\item{compare}{numeric value to compare against reference (base) value}

\item{reference}{numeric value that 'compare' value will be compared against}

\item{headline}{a string to format the final output. Uses
\code{\link[glue]{glue}} syntax}

\item{if_match}{string to display if numbers match, uses
\code{\link[glue]{glue}} syntax}

\item{trend_phrases}{list of values to use for when y is more than x, y is the
same as x, or y is less than x.}

\item{plural_phrases}{named list of values to use when difference (delta) is
singular (delta = 1) or plural (delta != 1)}

\item{orig_values}{a string to display the two original values. Uses
\code{\link[glue]{glue}} syntax. \code{{c}} = the 'compare' value, and
\code{{r}} = 'reference'}

\item{n_decimal}{numeric value to limit the number of decimal places in
the returned values.}

\item{round_all}{logical value to indicate if all values should be rounded.
When FALSE, the values will return with no modification. When TRUE (default)
all values will be round to the length specified by 'n_decimal'.}

\item{multiplier}{number indicating the scaling factor. When multiplier = 1
(default), 0.25 will return 0.25. When multiplier = 100, 0.25 will return 25.}

\item{return_data}{logical to indicate whether function should return the
phrase components used to compose the headline}

\item{x}{a list with values to compare, if named, can call by name}
}
\description{
Compose phrases that describe differences in the data
}
\examples{
# values can be manually entered, some headlines are provided by default
headline(10, 8)
headline(8, 10)
headline(10, 10)

# most likely you'll edit the headline by hand
headline(
  compare = 10,
  reference = 8,
  headline = "There was a ${delta} {trend} vs last year"
)

# you can also adjust the phrasing of higher/lower values
headline(
  compare = 10,
  reference = 8,
  headline = "Group A was {trend} by {delta_p}\%.",
  trend_phrases = trend_terms(more = "higher", less = "lower")
 )

# a phrase about the comparion can be edited by providing glue syntax
# 'c' = the 'compare' value, 'r' = 'reference'
headline(10, 8, orig_values = "{c} to {r} people")

# you can also add phrases for when the difference = 1 or not
headline(
  compare = 10,
  reference = 8,
  plural_phrases = list(
    were = plural_phrasing(single = "was", multi = "were"),
    people = plural_phrasing(single = "person", multi = "people")
  ),
  headline = "there {were} {delta} {people}"
)

# you can also adjust the rounding, although the default is 1
headline(0.1234, 0.4321)
headline(0.1234, 0.4321, n_decimal = 3)
# or use a multiplier
headline(0.1234, 0.4321, multiplier = 100)

# there are many components you can assemble
headline(
  compare = 16,
  reference = 8,
  headline = "there was {article_delta_p}\% {trend}, \\\\
  {article_trend} {trend} of {delta} ({orig_values})"
)

# compare_conditions() produces a list that can be passed to headline()
 mtcars \%>\%
   compare_conditions(
     compare = cyl == 4,
     reference = cyl == 6,
     cols = c(mpg)
   ) \%>\%
   headline(
     headline =
       "4-cylinder cars get an average of {delta} {trend} miles \\\\
       per gallon than 6-cylinder cars ({orig_values}).",
     trend_phrases = trend_terms("more", "less")
   )

}
\seealso{
\code{\link[=view_list]{view_list()}} and \code{\link[=trend_terms]{trend_terms()}}
}
