% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headline.R
\name{headline.default}
\alias{headline.default}
\title{Compare two values and get talking points}
\usage{
\method{headline}{default}(
  compare,
  reference,
  headline = "{trend} of {delta} ({orig_values})",
  ...,
  if_match = "There was no difference.",
  trend_phrasing = headliner::trend_terms(),
  orig_values = "{c} vs. {r}",
  n_decimal = 1,
  round_all = TRUE,
  scale = 100,
  return_data = FALSE
)
}
\arguments{
\item{compare}{numeric value to compare against reference (base) value}

\item{reference}{numeric value that 'compare' value will be compared against}

\item{headline}{a string to format the final output. Uses
\code{\link[glue]{glue}} syntax}

\item{if_match}{string to display if numbers match, uses
\code{\link[glue]{glue}} syntax}

\item{trend_phrasing}{list of values to use for when y is more than x, y is the
same as x, or y is less than x.}

\item{orig_values}{a string to display the two original values. Uses
\code{\link[glue]{glue}} syntax. \code{{c}} = the 'compare' value, and
\code{{r}} = 'reference'}

\item{n_decimal}{numeric value to limit the number of decimal places in
the returned values.}

\item{round_all}{logical value to indicate if all values should be rounded.
When FALSE, the values will return with no modification. When TRUE (default)
all values will be round to the length specified by 'n_decimal'.}

\item{scale}{number indicating the scaling factor. When scale = 1, 1/4 will
return 0.25, when scale = 100 (default) 1/4 will return 25}
}
\description{
Compare two values and get talking points
}
\examples{
# values can be manually entered, some headlines are provided by default
headline(10, 8)
headline(8, 10)
headline(10, 10)

# most likely you'll edit the headline by hand
headline(
  compare = 10,
  reference = 8,
  headline = "There was a ${delta} {trend} vs last year"
)

# you can also adjust the phrasing of higher/lower values
headline(
  compare = 10,
  reference = 8,
  headline = "Group A was {trend} by {delta_p}\%.",
  trend_phrasing = trend_terms(more = "higher", less = "lower")
 )

# a phrase about the comparion can be edited by providing glue syntax
# 'c' = the 'compare' value, 'r' = 'reference'
headline(10, 8, orig_values = "{c} to {r} people")

# you can also adjust the rounding, although the default is 1
headline(0.1234, 0.4321)
headline(0.1234, 0.4321, n_decimal = 3)

# The values can come from a summarized data frame or a named list
# if the data frame is only 2 columns or the list has only 2 elements
# you don't need to specify the 'compare' or 'reference' arguments unless
# you need to change the order. If the # of columns is > 3 you'll need to
# specify these arguments.
iris \%>\%
  dplyr::summarise_at(dplyr::vars(Sepal.Length, Petal.Length), mean) \%>\%
  headline()

iris \%>\%
  dplyr::summarise_at(dplyr::vars(Sepal.Length, Petal.Length), mean) \%>\%
  headline(
    compare = Petal.Length,
    reference = Sepal.Length
  )

# compare_conditions() produces a list that can be passed to headline()
 mtcars \%>\%
   compare_conditions(
     compare = cyl == 4,
     reference = cyl == 6,
     cols = c(mpg)
   ) \%>\%
   headline(
     headline =
       "4-cylinder cars get an average of {delta} {trend} miles \\\\
       per gallon than 6-cylinder cars ({orig_values}).",
     trend_phrasing = trend_terms("more", "less")
   )
}
\seealso{
\code{\link[=view_list]{view_list()}} and \code{\link[=trend_terms]{trend_terms()}}
}
