% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plural-phrasing.R
\name{plural_phrasing}
\alias{plural_phrasing}
\title{List of values to use when change is plural (or singlular)}
\usage{
plural_phrasing(single, multi)
}
\arguments{
\item{single}{string to use when delta = 1}

\item{multi}{string to use when delta > 1}
}
\description{
List of values to use when change is plural (or singlular)
}
\examples{
plural_phrasing("person", "people")
}
