% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-defaults.R
\name{set_headliner_defaults}
\alias{set_headliner_defaults}
\title{Set the default values used in making headlines}
\usage{
set_headliner_defaults(headline = NULL, orig_values = NULL)
}
\arguments{
\item{default}{string using component names from \code{\link[=compare_values]{compare_values()}}.
The default is \code{"{trend} of {delta} ({orig_values})"}}
}
\description{
Set the default values used in making headlines
}
\examples{
# the way headliner writes out-of-the-box
headline(3, 4)

# the values headliner uses under the hood
show_headliner_defaults()

# you can use compare_values(2, 3) \%>\% view_list() to see all talking
# points but we can use this as a short cut to see some of them:
names(compare_values(2, 3)) \%>\% head()

# update headline
set_headliner_defaults(
  headline = "{article_delta_p}\% {trend} ({orig_values})",
  orig_values = "from {r} to {c}"
)

# view changes
show_headliner_defaults()

# and in action
headline(3, 4)

# reset values
set_headliner_defaults(headline = NULL, orig_values = NULL)
headline(3, 4)
}
\seealso{
\code{\link[=augment_article_patterns]{augment_article_patterns()}}, \code{\link[=show_headliner_defaults]{show_headliner_defaults()}}
}
