% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-phrase.R
\name{trend_terms}
\alias{trend_terms}
\title{Phrases for direction of difference}
\usage{
trend_terms(more = "increase", less = "decrease", same = "difference")
}
\arguments{
\item{more}{string to use when x > y}

\item{less}{string to use when x < y}

\item{same}{string to use when x == y}
}
\description{
Phrases for direction of difference
}
\examples{
trend_terms(same = "no change")
}
\seealso{
\code{\link[=compare_values]{compare_values()}}
}
