% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-list.R
\name{view_list}
\alias{view_list}
\title{Compact view of list values}
\usage{
view_list(x)
}
\arguments{
\item{x}{a vector or list to be transposed}
}
\value{
Returns a data frame to display a list or vector vertically.
}
\description{
Compact view of list values
}
\examples{
compare_values(10, 8) |>
  view_list()

add_article(c(1,8,10, 11, 18)) |>
  view_list()
}
\seealso{
\code{\link[=compare_values]{compare_values()}}
}
